Adding repository for content table

+++ src/main/java/com/luiza/cms/entity/ContentEntity.java
@@ -9,7 +9,7 @@ public class ContentEntity {
 
     @Id
     @Column(name = "id", nullable = false)
-    private int id;
+    private Long id;
 
     @Column(name = "name", nullable = false)
     private String name;
@@ -30,13 +30,13 @@ public class ContentEntity {
     private String url;
 
     @Column(name = "parent_id")
-    private int parentId;
+    private Long parentId;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -88,11 +88,11 @@ public class ContentEntity {
         this.url = url;
     }
 
-    public int getParentId() {
+    public Long getParentId() {
         return parentId;
     }
 
-    public void setParentId(int parentId) {
+    public void setParentId(Long parentId) {
         this.parentId = parentId;
     }
 }
